Bksf (#123)
* Created the bksf file.

This file contains all the required methods for the BKSF algorithm. The structure is similar to _jordan_wigner.py.
It takes interactionoperator object as the input and returns the QubitOperator object

* Created _test_bksf.py file for the test routines for bksf algorithms

* Update _bksf.py

* Variable names, Args, Descrition of BKSF

Changed variable names to be little more intuitive. Added Args: in the docstring. Explanation for the BKSF algorithm has also been added. This would also help in making sense of those long expressions.

* Assert statement instead of printing

Fixed the file name. Removed all the commented out lines. Now, while comparing the spectrum from Jordan-Wigner and BKSF I am asserting that half the spectrum should be same.

* Changing the import statement

* Added networkx to the requirements

* Fixing the filename

* AssertTrue changed to AssertEqual

* Fixed the division error

* Test for generate_fermions()

* some cosmetic changes

* More tests

Added more functional tests. Writing unit tests for each of the module is hard. But, the functional tests use different portions of the code, so, debugging should be easier if an error were to occur.

* Increased test coverage

* fixing import error

* fixing the error that reduced test coverage